--- src/pages/HomePage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import HomeContainer from "../containers/home/HomeContainer";
-
-const HomePage = () => {
-  return <HomeContainer />;
-};
-
-export default HomePage;
