--- src/pages/LoginPage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import LoginContainer from "../containers/auth/LoginContainer";
-
-const LoginPage = () => {
-  return <LoginContainer />;
-};
-
-export default LoginPage;
