--- src/pages/SignupPage.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import SingupContainer from "../containers/auth/SignupContainer";
-
-const SignupPage = () => {
-  return <SingupContainer />;
-};
-
-export default SignupPage;
